+++ src/services/InstagramService.php
@@ -11,10 +11,16 @@ use craft\errors\InvalidVolumeException;
 use craft\helpers\FileHelper;
 use Exception;
 use GuzzleHttp\Client;
+use GuzzleHttp\Cookie\CookieJar;
 use GuzzleHttp\Exception\GuzzleException;
+use yii\caching\TagDependency;
 
 class InstagramService extends Component
 {
+    public const CACHE_TAG = 'instagramfeed';
+
+    private const DEFAULT_USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/77.0.3865.75 Safari/537.36';
+
     /**
      * Returns the current Instagram feed of the configured account.
      *
@@ -36,6 +42,11 @@ class InstagramService extends Component
          */
         $settings = InstagramFeed::getInstance()->getSettings();
 
+        // Cache and cache tag
+        $cacheService = Craft::$app->getCache();
+        $dependency = new TagDependency();
+        $dependency->tags[] = self::CACHE_TAG;
+
         if ($accountOrTag === null) {
             $accountOrTag = $settings->getAccount();
             if (empty($accountOrTag)) {
@@ -50,9 +61,9 @@ class InstagramService extends Component
         $accountOrTag = strtolower($accountOrTag);
         $hash = md5($settings->useVolume . $settings->volume . $settings->subpath . $accountOrTag);
 
-        $uptodate = Craft::$app->getCache()->get('instagram_uptodate_' . $hash);
-        $cachedItems = Craft::$app->getCache()->get('instagram_data_' . $hash);
-        $timeout = Craft::$app->getCache()->get('instagram_update_error_' . $hash);
+        $uptodate = $cacheService->get('instagram_uptodate_' . $hash);
+        $cachedItems = $cacheService->get('instagram_data_' . $hash);
+        $timeout = $cacheService->get('instagram_update_error_' . $hash);
 
         if ($uptodate === false && $timeout === false) {
             Craft::debug('No cached data found, start fetching Instagram page.', __METHOD__);
@@ -67,8 +78,9 @@ class InstagramService extends Component
 
             if (!empty($items)) {
                 Craft::debug('Items found, caching them.', __METHOD__);
-                Craft::$app->getCache()->set('instagram_data_' . $hash, $items, 2592000);
-                Craft::$app->getCache()->set('instagram_uptodate_' . $hash, true, 21600);
+                $cacheService->set('instagram_data_' . $hash, $items, 2592000, $dependency);
+                $cacheService->set('instagram_uptodate_' . $hash, true, 21600, $dependency);
+                $cacheService->set('instagram_update_error_' . $hash, false, 0, $dependency);
 
                 return $this->populateImages($items);
             }
@@ -76,16 +88,16 @@ class InstagramService extends Component
             if (!empty($cachedItems)) {
                 // If not updated expand cache time and set update to six hours to stop from retrying every request
                 Craft::debug('Error fetching new data from Instagram, using existing cached data and expanding cache time. Stopping requests for 15 minutes.', __METHOD__);
-                Craft::$app->getCache()->set('instagram_data_' . $hash, $cachedItems, 2592000);
-                Craft::$app->getCache()->set('instagram_update_error_' . $hash, true, 21600);
+                $cacheService->set('instagram_data_' . $hash, $cachedItems, 2592000, $dependency);
+                $cacheService->set('instagram_update_error_' . $hash, true, 21600, $dependency);
             }
 
             if (empty($cachedItems)) {
                 // If the cache is empty (e.g. first request ever) and the request fails, we are stopping requests for some time.
                 $waitTime = $this->canUseProxy() ? 10 : 900;
                 Craft::debug('Cache is empty and no items could be fetched. Stopping requests for ' . $waitTime . ' seconds.', __METHOD__);
-                Craft::$app->getCache()->set('instagram_data_' . $hash, [], 2592000);
-                Craft::$app->getCache()->set('instagram_update_error_' . $hash, true, $waitTime);
+                $cacheService->set('instagram_data_' . $hash, [], 2592000, $dependency);
+                $cacheService->set('instagram_update_error_' . $hash, true, $waitTime, $dependency);
 
                 return [];
             }
@@ -100,32 +112,20 @@ class InstagramService extends Component
      * Fetches the feed from the public Instagram profile page.
      *
      * @param string $account The account name to fetch.
-     *
      * @return array
      * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \craft\errors\SiteNotFoundException
      */
     private function getInstagramAccountData(string $account): array
     {
-        $html = $this->fetchInstagramPage($account . '/');
+        $html = $this->fetchInstagramAccount($account);
 
         if (null === $html) {
-            Craft::warning('Instagram profile data could not be fetched.', __METHOD__);
+            Craft::error('Instagram profile data could not be fetched.', 'instagramfeed');
 
             return [];
         }
 
-        if ($this->canUseProxy()) {
-            $obj = $this->parseProxyResponse($html);
-            if (false === $obj) {
-                return [];
-            }
-        } else {
-            $obj = $this->parseInstagramResponse($html);
-            if (empty($obj)) {
-                return [];
-            }
-        }
+        $obj = $this->parseResponse($html);
 
         return $this->extractMedia($obj);
     }
@@ -134,87 +134,124 @@ class InstagramService extends Component
      * Fetches the feed from the public Instagram tag page.
      *
      * @param string $tag The tag name to fetch.
-     *
      * @return array
      * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \craft\errors\SiteNotFoundException
      */
     private function getInstagramTagData(string $tag): array
     {
-        $tag = substr($tag, 1);
-
-        $path = sprintf('explore/tags/%s/', $tag);
-        $html = $this->fetchInstagramPage($path);
+        $html = $this->fetchInstagramTag($tag);
 
         if (null === $html) {
-            Craft::error('Instagram tag data could not be fetched.', __METHOD__);
+            Craft::error('Instagram tag data could not be fetched.', 'instagramfeed');
 
             return [];
         }
 
-        if ($this->canUseProxy()) {
-            $obj = $this->parseProxyResponse($html);
-            if (false === $obj) {
-                return [];
-            }
-        } else {
-            $obj = $this->parseInstagramResponse($html);
-            if (empty($obj)) {
-                return [];
-            }
-        }
+        $obj = $this->parseResponse($html);
 
         return $this->extractMedia($obj);
     }
 
     /**
-     * Fetches the page from a given URL
+     * Fetches an account from Instagram.
      *
-     * @param string $path The path to fetch
-     *
-     * @return false|string
+     * @param string $account The account name to fetch.
+     * @return string|null
      * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \craft\errors\SiteNotFoundException
      */
-    private function fetchInstagramPage(string $path): ?string
+    private function fetchInstagramAccount(string $account): ?string
     {
-        /**
-         * @var \codemonauts\instagramfeed\models\Settings $settings
-         */
         $settings = InstagramFeed::getInstance()->getSettings();
-
-        $defaultUserAgent = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/77.0.3865.75 Safari/537.36';
-        $url = 'https://www.instagram.com/' . $path;
-
-        $client = new Client();
+        $cookies = new CookieJar();
+        $client = new Client([
+            'cookies' => $cookies,
+        ]);
 
         $guzzleOptions = [
             'timeout' => $settings->timeout,
             'headers' => [
                 'Accept-Language' => 'en-US;q=0.9,en;q=0.8',
-                'User-Agent' => $settings->userAgent !== '' ? $settings->userAgent : $defaultUserAgent,
+                'User-Agent' => self::DEFAULT_USER_AGENT,
             ],
         ];
 
-        if ($this->canUseProxy()) {
-            $referer = Craft::$app->getSites()->getCurrentSite()->getBaseUrl();
-            $url = 'https://igproxy.codemonauts.com/' . $path;
-
-            $guzzleOptions['headers']['Authorization'] = $settings->proxyKey;
-            $guzzleOptions['headers']['Referer'] = $referer;
-            $guzzleOptions['headers']['User-Agent'] = $defaultUserAgent;
-            $guzzleOptions['headers']['X-Plugin-Version'] = InstagramFeed::getInstance()->getVersion();
-        }
-
         try {
-            $response = $client->get($url, $guzzleOptions);
+            if ($this->canUseProxy()) {
+                $url = sprintf('https://igproxy.codemonauts.com/%s/', $account);
+                $referer = Craft::$app->getSites()->getCurrentSite()->getBaseUrl();
+                $guzzleOptions['headers']['Authorization'] = $settings->proxyKey;
+                $guzzleOptions['headers']['Referer'] = $referer;
+                $guzzleOptions['headers']['X-Plugin-Version'] = InstagramFeed::getInstance()->getVersion();
+                $response = $client->get($url, $guzzleOptions);
+            } else {
+                $url = sprintf('https://www.instagram.com/%s/', $account);
+                $client->get($url, $guzzleOptions);
+                $guzzleOptions['headers']['Referer'] = $url;
+                $guzzleOptions['headers']['X-IG-App-ID'] = '936619743392459';
+                $guzzleOptions['headers']['X-Requested-With'] = 'XMLHttpRequest';
+                $guzzleOptions['headers']['Origin'] = 'https://www.instagram.com';
+                $url = sprintf('https://www.instagram.com/api/v1/users/web_profile_info/?username=%s', $account);
+                $response = $client->get($url, $guzzleOptions);
+            }
         } catch (Exception $e) {
-            Craft::error($e->getMessage(), __METHOD__);
+            Craft::error('Error fetching page: ' . $e->getMessage(), 'instagramfeed');
 
             return null;
         }
 
-        return $response->getBody();
+        return (string)$response->getBody();
+    }
+
+    /**
+     * Fetches a tag from Instagram.
+     *
+     * @param string $tag The tag to fetch.
+     * @return string|null
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     */
+    private function fetchInstagramTag(string $tag): ?string
+    {
+        $settings = InstagramFeed::getInstance()->getSettings();
+        $tag = substr($tag, 1);
+
+        $cookies = new CookieJar();
+        $client = new Client([
+            'cookies' => $cookies,
+        ]);
+
+        $guzzleOptions = [
+            'timeout' => $settings->timeout,
+            'headers' => [
+                'Accept-Language' => 'en-US;q=0.9,en;q=0.8',
+                'User-Agent' => self::DEFAULT_USER_AGENT,
+            ],
+        ];
+
+        try {
+            if ($this->canUseProxy()) {
+                $url = sprintf('https://igproxy.codemonauts.com/explore/tags/%s/', $tag);
+                $referer = Craft::$app->getSites()->getCurrentSite()->getBaseUrl();
+                $guzzleOptions['headers']['Authorization'] = $settings->proxyKey;
+                $guzzleOptions['headers']['Referer'] = $referer;
+                $guzzleOptions['headers']['X-Plugin-Version'] = InstagramFeed::getInstance()->getVersion();
+                $response = $client->get($url, $guzzleOptions);
+            } else {
+                $url = sprintf('https://www.instagram.com/explore/tags/%s/', $tag);
+                $client->get($url, $guzzleOptions);
+                $guzzleOptions['headers']['Referer'] = $url;
+                $guzzleOptions['headers']['X-IG-App-ID'] = '936619743392459';
+                $guzzleOptions['headers']['X-Requested-With'] = 'XMLHttpRequest';
+                $guzzleOptions['headers']['Origin'] = 'https://www.instagram.com';
+                $url = sprintf('https://www.instagram.com/api/v1/tags/logged_out_web_info/?tag_name=%s', $tag);
+                $response = $client->get($url, $guzzleOptions);
+            }
+        } catch (Exception $e) {
+            Craft::error('Error fetching page: ' . $e->getMessage(), 'instagramfeed');
+
+            return null;
+        }
+
+        return (string)$response->getBody();
     }
 
     /**
@@ -224,7 +261,7 @@ class InstagramService extends Component
      *
      * @return mixed
      */
-    private function parseProxyResponse(string $response)
+    private function parseResponse(string $response)
     {
         /**
          * @var \codemonauts\instagramfeed\models\Settings $settings
@@ -235,44 +272,12 @@ class InstagramService extends Component
             $this->dumpResponse($response);
         }
 
-        return json_decode($response, true);
-    }
-
-    /**
-     * Function to parse the response body from Instagram
-     *
-     * @param string $response Response body from Instagram
-     *
-     * @return array|boolean
-     */
-    private function parseInstagramResponse(string $response)
-    {
-        /**
-         * @var \codemonauts\instagramfeed\models\Settings $settings
-         */
-        $settings = InstagramFeed::getInstance()->getSettings();
-
-        $arr = explode('window._sharedData = ', $response);
-
-        if (!isset($arr[1])) {
-            // Check if Instagram returned a statement and not a valid page
-            $statement = json_decode($response, false);
-            if (isset($statement->errors)) {
-                Craft::error('Instagram responsed with an error: ' . implode(' ', $statement->errors->error), 'instagramfeed');
-            } else {
-                Craft::error('Unknown response from Instagram. Please check debug output in devMode.', 'instagramfeed');
-            }
-
-            $this->dumpResponse($response);
+        $data = json_decode($response, true);
+        if (false === $data || null === $data) {
             return [];
         }
 
-        if ($settings->dump) {
-            $this->dumpResponse($response);
-        }
-
-        $arr = explode(';</script>', $arr[1]);
-        return json_decode($arr[0], true);
+        return $data;
     }
 
     /**