+++ src/structures.php
@@ -6,8 +6,16 @@ return [
         'parser' => codemonauts\instagramfeed\parsers\AccountVersion1Parser::class,
     ],
     [
+        'structure' => 'items.0',
+        'parser' => codemonauts\instagramfeed\parsers\AccountVersion2Parser::class,
+    ],
+    [
         'structure' => 'entry_data.TagPage.0.graphql.hashtag.edge_hashtag_to_media.edges',
         'parser' => codemonauts\instagramfeed\parsers\TagVersion1Parser::class,
+    ],
+    [
+        'structure' => 'data.top.sections',
+        'parser' => codemonauts\instagramfeed\parsers\TagVersion2Parser::class,
     ],
     [
         'structure' => 'data.recent.sections',